requirements.txt
# Core AI and ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Cybersecurity and networking
scapy>=2.5.0
python-nmap>=0.7.1
paramiko>=3.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web application security
selenium>=4.15.0
webdriver-manager>=4.0.0

# System and process management
psutil>=5.9.0
subprocess32>=3.5.4

# Data processing and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Additional security tools
cryptography>=41.0.0
dnspython>=2.3.0
python-whois>=0.8.0 